# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Law OA Go API
  description: |
    Law OA Go 提供现代化的 RESTful API，基于 Go 1.23+ 标准构建，采用统一响应格式、类型安全处理和完善的中间件体系。
    
    ## 核心特性
    - ✅ **统一响应格式**: 所有API使用一致的响应结构
    - ✅ **类型安全**: 基于 Go generics 的类型安全处理
    - ✅ **现代中间件**: CORS、限流、JWT认证、超时控制
    - ✅ **完善错误处理**: 标准化错误响应和建议信息
    - ✅ **请求追踪**: 每个请求都有唯一ID用于调试
    - ✅ **分页支持**: 统一的分页参数和响应格式
    
    ## 基础信息
    - **Base URL**: http://localhost:8080/api/v1
    - **认证方式**: Bearer Token (JWT)
    - **数据格式**: JSON
    - **API版本**: v1
    - **Go版本**: 1.23+
  version: 1.0.0
  contact:
    name: Law OA Go Team
    email: support@lawfirm.com
    url: https://github.com/law-oa-go
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发服务器
  - url: https://api.lawfirm.com/v1
    description: 生产服务器
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证令牌，通过登录接口获取
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 请求是否成功
        meta:
          $ref: '#/components/schemas/Meta'
      required:
        - success
        - meta

    # 成功响应
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              description: 响应数据
          required:
            - data

    # 分页响应
    PageResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              description: 数据数组
              items:
                description: 数据项
            pagination:
              $ref: '#/components/schemas/Pagination'
          required:
            - data
            - pagination

    # 错误响应
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              $ref: '#/components/schemas/ErrorDetail'
          required:
            - error

    # 错误详情
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: 错误码
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: 错误消息
          example: "Validation failed"
        details:
          type: string
          description: 错误详情
          example: "Email format is invalid"
        suggestions:
          type: array
          description: 建议解决方案
          items:
            type: string
          example: ["Check email format"]
      required:
        - code
        - message

    # 元数据
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
          example: "2024-01-01T00:00:00Z"
        request_id:
          type: string
          description: 请求ID
          example: "req_123456"
        version:
          type: string
          description: API版本
          example: "1.0.0"
      required:
        - timestamp
        - request_id
        - version

    # 分页信息
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: 当前页码
          example: 1
        page_size:
          type: integer
          description: 每页数量
          example: 20
        total:
          type: integer
          description: 总记录数
          example: 100
        total_pages:
          type: integer
          description: 总页数
          example: 5
      required:
        - page
        - page_size
        - total
        - total_pages

    # 用户模型
    User:
      type: object
      properties:
        id:
          type: integer
          description: 用户ID
          example: 1
        name:
          type: string
          description: 用户姓名
          example: "系统管理员"
        email:
          type: string
          format: email
          description: 邮箱地址
          example: "admin@lawfirm.com"
        role:
          type: string
          description: 用户角色
          enum: [admin, user, lawyer]
          example: "admin"
        phone:
          type: string
          description: 电话号码
          example: "13800138000"
        avatar:
          type: string
          description: 头像URL
          example: "https://example.com/avatar.jpg"
        status:
          type: string
          description: 用户状态
          enum: [active, inactive, suspended]
          example: "active"
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: "2024-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: "2024-01-01T00:00:00Z"
      required:
        - id
        - name
        - email
        - role
        - status
        - created_at
        - updated_at

    # 客户模型
...
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          description: 用户姓名
          example: "新用户"
        email:
          type: string
          format: email
          description: 邮箱地址
          example: "newuser@example.com"
        password:
          type: string
          format: password
          description: 密码
          example: "password123"
        phone:
          type: string
          description: 电话号码
          example: "13800138000"
        role:
          type: string
          description: 用户角色
          enum: [admin, user, lawyer]
          example: "user"
      required:
        - name
        - email
        - password
        - phone
        - role

    # 创建客户请求
...
paths:
...
  /admin/users:
    get:
      tags:
        - 用户管理
      summary: 获取用户列表
      description: 获取系统用户列表，支持分页和搜索
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: 每页数量
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: 搜索关键词
          schema:
            type: string
        - name: role
          in: query
          description: 角色过滤
          schema:
            type: string
            enum: [admin, user, lawyer]
        - name: status
          in: query
          description: 状态过滤
          schema:
            type: string
            enum: [active, inactive, suspended]
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - 用户管理
      summary: 创建用户
      description: 创建新用户账户
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RegisterRequest'
                - type: object
                  properties:
                    status:
                      type: string
                      description: 用户状态
                      enum: [active, inactive, suspended]
                      default: "active"
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 用户已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{id}:
    get:
      tags:
        - 用户管理
      summary: 获取用户详情
      description: 根据ID获取用户详细信息
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 用户ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - 用户管理
      summary: 更新用户
      description: 更新用户信息
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 用户ID
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 用户姓名
                email:
                  type: string
                  format: email
                  description: 邮箱地址
                phone:
                  type: string
                  description: 电话号码
                role:
                  type: string
                  description: 用户角色
                  enum: [admin, user, lawyer]
                status:
                  type: string
                  description: 用户状态
                  enum: [active, inactive, suspended]
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 邮箱已被使用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - 用户管理
      summary: 删除用户
      description: 删除指定用户
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 用户ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "User deleted successfully"
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 客户管理接口
...