# internal/handlers/user_handler.go
package handlers

import (
	"github.com/gin-gonic/gin"
	"law-oa-go/internal/services"
)

type UserHandler struct {
	userService *services.UserService
}

func NewUserHandler(userService *services.UserService) *UserHandler {
	return &UserHandler{
		userService: userService,
	}
}

// ListUsers godoc
// @Summary 获取用户列表
// @Description 分页获取用户列表，支持过滤和搜索
// @Tags 用户管理
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param page query int false "页码" default(1)
// @Param page_size query int false "每页数量" default(20)
// @Param role query string false "用户角色" Enums(admin,lawyer,user)
// @Param status query string false "用户状态" Enums(active,inactive)
// @Param search query string false "搜索关键词"
// @Success 200 {object} common.APIResponse{data=[]services.UserProfile} "获取成功"
// @Failure 400 {object} common.APIResponse "请求参数错误"
// @Failure 401 {object} common.APIResponse "未授权"
// @Failure 403 {object} common.APIResponse "权限不足"
// @Failure 500 {object} common.APIResponse "内部错误"
// @Router /admin/users [get]
func (h *UserHandler) ListUsers(c *gin.Context) {
	handler := APIListHandler(func(c *gin.Context, req *services.UserListRequest) ([]*services.UserProfile, int64, error) {
		return h.userService.ListUsers(c.Request.Context(), req)
	}, "users")
	handler(c)
}

// GetUser godoc
// @Summary 获取用户详情
// @Description 根据ID获取用户详细信息
// @Tags 用户管理
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "用户ID"
// @Success 200 {object} common.APIResponse{data=services.UserProfile} "获取成功"
// @Failure 400 {object} common.APIResponse "请求参数错误"
// @Failure 401 {object} common.APIResponse "未授权"
// @Failure 403 {object} common.APIResponse "权限不足"
// @Failure 404 {object} common.APIResponse "用户不存在"
// @Failure 500 {object} common.APIResponse "内部错误"
// @Router /admin/users/{id} [get]
func (h *UserHandler) GetUser(c *gin.Context) {
	handler := APIGetHandler(func(c *gin.Context, id uint) (*services.UserProfile, error) {
		return h.userService.GetUserProfile(c.Request.Context(), id)
	}, "user")
	handler(c)
}

// CreateUser godoc
// @Summary 创建用户
// @Description 创建新用户账户
// @Tags 用户管理
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body services.CreateUserRequest true "创建用户请求"
// @Success 200 {object} common.APIResponse{data=services.UserProfile} "创建成功"
// @Failure 400 {object} common.APIResponse "请求参数错误"
// @Failure 401 {object} common.APIResponse "未授权"
// @Failure 403 {object} common.APIResponse "权限不足"
// @Failure 500 {object} common.APIResponse "内部错误"
// @Router /admin/users [post]
func (h *UserHandler) CreateUser(c *gin.Context) {
	handler := APICreateHandler(func(c *gin.Context, req *services.CreateUserRequest) (*services.UserProfile, error) {
		return h.userService.CreateUser(c.Request.Context(), req)
	}, "user")
	handler(c)
}

// UpdateUser godoc
// @Summary 更新用户信息
// @Description 更新指定用户的信息
// @Tags 用户管理
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "用户ID"
// @Param request body services.UpdateUserRequest true "更新用户请求"
// @Success 200 {object} common.APIResponse{data=services.UserProfile} "更新成功"
// @Failure 400 {object} common.APIResponse "请求参数错误"
// @Failure 401 {object} common.APIResponse "未授权"
// @Failure 403 {object} common.APIResponse "权限不足"
// @Failure 404 {object} common.APIResponse "用户不存在"
// @Failure 500 {object} common.APIResponse "内部错误"
// @Router /admin/users/{id} [put]
func (h *UserHandler) UpdateUser(c *gin.Context) {
	handler := APIUpdateHandler(func(c *gin.Context, id uint, req *services.UpdateUserRequest) (*services.UserProfile, error) {
		return h.userService.UpdateUser(c.Request.Context(), id, req)
	}, "user")
	handler(c)
}

// DeleteUser godoc
// @Summary 删除用户
// @Description 删除指定的用户
// @Tags 用户管理
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "用户ID"
// @Success 200 {object} common.APIResponse "删除成功"
// @Failure 400 {object} common.APIResponse "请求参数错误"
// @Failure 401 {object} common.APIResponse "未授权"
// @Failure 403 {object} common.APIResponse "权限不足"
// @Failure 404 {object} common.APIResponse "用户不存在"
// @Failure 500 {object} common.APIResponse "内部错误"
// @Router /admin/users/{id} [delete]
func (h *UserHandler) DeleteUser(c *gin.Context) {
	handler := APIDeleteHandler(func(c *gin.Context, id uint) error {
		return h.userService.DeleteUser(c.Request.Context(), id)
	}, "user")
	handler(c)
}